Work in progress; Refactor user model, endpoints

+++ FitnessServer/src/main/java/hu/bme/fitnessapplication/server/endpoint/v1/user/UserController.java
@@ -1,6 +1,7 @@
-package hu.bme.fitnessapplication.server.endpoint.user;
+package hu.bme.fitnessapplication.server.endpoint.v1.user;
 
 import hu.bme.fitnessapplication.server.BaseService;
+import hu.bme.fitnessapplication.server.endpoint.v1.BaseController;
 import hu.bme.fitnessapplication.server.repository.user.model.Role;
 import hu.bme.fitnessapplication.server.repository.user.model.User;
 import hu.bme.fitnessapplication.server.repository.user.model.dto.UserRequestDTO;
@@ -10,28 +11,30 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
-@RestController
-@RequestMapping("/user")
-public class UserController {
+@RestController("UserController_v1")
+@RequestMapping("v1/users")
+public class UserController extends BaseController {
 
     @Autowired
     private UserService userService;
 
     /**
      * Returns the currently logged in user
+     *
      * @return
      */
     @RequestMapping(path = "/me", method = RequestMethod.GET)
-    public ResponseEntity getMe() {
-        User me = userService.getLoggedInUser();
+    public ResponseEntity getMe(Authentication auth) {
+        User me = getCallingUser(auth);
         if (me != null) {
-            return ResponseEntity.ok(new UserResponseDTO(me));
+            return ResponseEntity.ok(convertToDto(me));
         } else {
             return ResponseEntity.notFound().build();
         }
@@ -40,14 +43,14 @@ public class UserController {
     @RequestMapping(method = RequestMethod.GET)
     @Secured("ROLE_ADMIN")
     public ResponseEntity getAllUsers() {
-       List<User> users = userService.findAll();
-       List<UserResponseDTO> result = new ArrayList<>();
+        List<User> users = userService.findAll();
+        List<UserResponseDTO> result = new ArrayList<>();
 
-       for (User user : users) {
-           result.add(new UserResponseDTO(user));
-       }
+        for (User user : users) {
+            result.add(convertToDto(user));
+        }
 
-       return ResponseEntity.ok(result);
+        return ResponseEntity.ok(result);
     }
 
 
@@ -59,7 +62,7 @@ public class UserController {
             User user = userService.findById(id);
 
             if (user != null) {
-                return ResponseEntity.ok(new UserResponseDTO(user));
+                return ResponseEntity.ok(convertToDto(user));
             } else {
                 return ResponseEntity.notFound().build();
             }
@@ -70,14 +73,13 @@ public class UserController {
     }
 
     @RequestMapping(method = RequestMethod.POST)
-    @Secured("ROLE_ADMIN")
     public ResponseEntity registerUser(@RequestBody UserRequestDTO userDTO) {
         if (userDTO == null) {
             return ResponseEntity.badRequest().build();
         }
 
         try {
-            User user = userDTO.toEntity();
+            User user = convertToEntity(userDTO);
             user = userService.create(user);
             return ResponseEntity.ok(new UserResponseDTO(user));
         } catch (UserService.UsernameAlreadyTakenException e) {
@@ -88,17 +90,16 @@ public class UserController {
     }
 
     @RequestMapping(path = "/{userId}", method = RequestMethod.PUT)
-    @Secured("ROLE_ADMIN")
-    public ResponseEntity updateUser(@PathVariable String userId, @RequestBody UserRequestDTO userDTO) {
-        if (userDTO == null) {
+    public ResponseEntity updateUser(@PathVariable String userId, @RequestBody UserRequestDTO userDto) {
+        if (userDto == null) {
             return ResponseEntity.badRequest().build();
         }
 
         try {
             UUID id = UUID.fromString(userId);
-            User updatedUser = userDTO.toEntity();
+            User updatedUser = convertToEntity(userDto);
             updatedUser = userService.update(id, updatedUser);
-            return ResponseEntity.ok(new UserResponseDTO(updatedUser));
+            return ResponseEntity.ok(convertToDto(updatedUser));
         } catch (BaseService.EntityNotFoundException e) {
             return ResponseEntity.notFound().build();
         } catch (BaseService.InvalidEntityException | IllegalArgumentException e) {
@@ -107,53 +108,58 @@ public class UserController {
     }
 
     @RequestMapping(path = "/{userId}", method = RequestMethod.DELETE)
-    @Secured("ROLE_ADMIN")
-    public ResponseEntity deleteUser(@PathVariable String userId) {
-        try {
+    public ResponseEntity deleteUser(@PathVariable String userId, Authentication auth) {
+        User callingUser = getCallingUser(auth);
+        if (callingUser != null) {
             UUID id = UUID.fromString(userId);
-            userService.delete(id);
+            if (callingUser.getId().equals(id) ||
+                    callingUser.getRole().getRole() == Role.ROLE_ADMIN) {
+                userService.delete(id);
+            }
             return ResponseEntity.ok().build();
-        } catch (BaseService.EntityNotFoundException e) {
-            return ResponseEntity.notFound().build();
+        } else {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
         }
     }
-    
+
     @RequestMapping(path = "/gyms", method = RequestMethod.GET)
     public ResponseEntity getGyms() {
-    	List<User> users = userService.findAll();
+        List<User> users = userService.findAll();
         List<UserResponseDTO> result = new ArrayList<>();
 
         for (User user : users) {
-        	if(user.getRole().getRole().name().equals(Role.ROLE_GYM))
-        		result.add(new UserResponseDTO(user));
+            if (user.getRole().getRole().name().equals(Role.ROLE_GYM))
+                result.add(new UserResponseDTO(user));
         }
 
         return ResponseEntity.ok(result);
     }
-    
-    @RequestMapping(path = "/trainers", method = RequestMethod.GET)
-    public ResponseEntity getTrainers() {
-    	List<User> users = userService.findAll();
-        List<UserResponseDTO> result = new ArrayList<>();
 
-        for (User user : users) {
-        	if(user.getRole().getRole().name().equals(Role.ROLE_TRAINER))
-        		result.add(new UserResponseDTO(user));
-        }
-
-        return ResponseEntity.ok(result);
-    }
-    
     @RequestMapping(path = "/users", method = RequestMethod.GET)
     public ResponseEntity getUsers() {
-    	List<User> users = userService.findAll();
+        List<User> users = userService.findAll();
         List<UserResponseDTO> result = new ArrayList<>();
 
         for (User user : users) {
-        	if(user.getRole().getRole().name().equals(Role.ROLE_USER))
-        		result.add(new UserResponseDTO(user));
+            if (user.getRole().getRole().name().equals(Role.ROLE_USER))
+                result.add(new UserResponseDTO(user));
         }
 
         return ResponseEntity.ok(result);
+    }
+
+    private User convertToEntity(UserRequestDTO userReq) {
+        User user = modelMapper.map(userReq, User.class);
+        return user;
+    }
+
+    private UserResponseDTO convertToDto(User user) {
+        UserResponseDTO dto = modelMapper.map(user, UserResponseDTO.class);
+        dto.setRole(user.getRole().getRole().name());
+        return dto;
+    }
+
+    protected User getCallingUser(Authentication auth) {
+        return userService.findByUsername(auth.getName());
     }
 }
